added delayed routing client
